requirements.txt
asgiref==3.8.1
astroid==3.3.8
attrs==25.3.0
click==8.1.8
colorama==0.4.6
coverage==7.7.0
dill==0.3.9
Django==5.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
h11==0.14.0
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
platformdirs==4.3.6
pylint==3.3.6
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
setuptools==65.5.0
sqlparse==0.5.3
tomlkit==0.13.2
tzdata==2025.1
uritemplate==4.1.1
uvicorn==0.34.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
